pyproject.toml
[project]
name = "ai-recipes"
version = "0.1.0"
description = "Extract recipes from images, web pages, and PDFs using AI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Text Processing :: General",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pypdf2>=3.0.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Platform-specific clipboard support
macos = ["pyobjc-framework-Cocoa>=9.0"]
windows = []  # Pillow already includes Windows clipboard support
linux = []  # Uses system tools (xclip or wl-paste)

[project.scripts]
ai-recipes = "src.cli:cli"

[project.urls]
Homepage = "https://github.com/yourusername/ai-recipes"
Repository = "https://github.com/yourusername/ai-recipes"
Issues = "https://github.com/yourusername/ai-recipes/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
]

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[dependency-groups]
dev = [
    "ruff>=0.12.7",
]
